Inventory HUD Upgrade
Changed the code in DrawInventoryHUD to have a similar HUD to minecraft. However, I had to create a new sprite for the still torch and make noteSprite and keySprite global variables in main.cpp but still have the textures drawn in there. gameutils.cpp will reference the variables from main.cpp